Fix for issue #98, adding esp_raincoast dataset (#107)
- Reproduced issue #98  with a test in `test_concat.py`
- To fix ^^, added a `ignore_index=True` parameter that resets the index
after operations like `drop_na`, `drop_duplicates` etc. In `filter.py`
we explicitly call `.reset_index(drop=True)` to do the same thing. But
we should discuss if this is the right way ? Currently the "index" of
any esp-data dataset doesnt really carry any information so its fine to
reset, but maybe it would be for certain use cases ? should we add info
on this as a warning to the Docs ?
- Added `esp_raincoast` dataset. This is a v1, might require a second PR
to fix some things as required by Emmanueal and co.