removing cloudpathlib (#123)
This PR removes cloudpathlib and just uses filesystem.

The main complexity is that fsspec doesn't work nicely with Path-like
objects. There has been quite a few attempts in the community to add
support to fsspec but it's complicated.

But we quite like Path-like objects so this PR extends PurePaths from
python to work with cloud Paths. This is not pretty but let's us at
least do foo / bar style operations and we explicitly use filesystem or
util functions that use filesystems for io operations.

- The way were processing and saving "data_root" made certain
assumptions and tests started failing after the changes in the `io`
module. Fixed a few issues when while I was at it:
- Noticed that the default data_root for audioset dataset was wrong.
Fixed and updated the tests. @GaganNarula we probably need new tests to
make sure that custom data_root works
- also noticed that some of data_root args in `__init__` had the old
style type hints. Have fixed some of those in this PR too.