Make cloud paths compatible with Python 3.12+ (#150)
Turns out relying on Python's internal APIs and implementation (pathlib
in particular) is not a great idea. :(

Python 3.12 did a big rehaul of pathlib implementation which broke the
interoperability with out PureCloudPath. I initially
[tried](892a7a3ec529e9604f902af911241ded95437068) to make it work and do
things differently based on Python version but that seems like something
that will need patching again soon.

Instead I de-coupled PureCloudPath from PurePath and made it its own
independant thing. There's still interoperability with Path/PurePath
(for example when joining with `/`).

I also revamped the unit tests (and moved them into `io` dir) with
Claude's help to cover more cases.

<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":"main"}
```
-->