Add python matrix version for CI subset (#152)
esp-data is a library and we don't control what python version our users
are going to use (besides forcing 3.10+). It's useful to test core
esp-data functionality with multiple python versions. This PR adds the
pipeline for doing that.

For now I've only added tests for `io` module into this step but
obviously we'll add more as we go.

@GaganNarula Adding this also exposed a race condition in our tests. We
have to be careful with paths in `@pytest.mark.parametrize` because if
they're fixed then running multiple tests at once (which this PR does)
can cause issues. I have fixed it in the io tests but we use a similar
pattern in other places.

<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":"main"}
```
-->