Added codeowners (#159)
Now that more people are making PRs I want to change ruleset so that
you'd need a review from Gagan/me.


<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":"main"}
```
-->