add iNaturalist dataset (#144)
Adding the iNaturalist dataset as a separate dataset in esp-data.

This points to a new version downloaded in mid-2025. It uses the
original sample rates (unlike the version available via the
"AnimalSpeak" dataset.) It includes a copy resampled in advance to
32khz, which will be used if the desired sampling rate matches
(specified via "available_sample_rates".)

Versioning for the future is still an open question for the moment, as
this dataset will be routinely updated. However, the path to the
metadata is delimited as
gs://esp-ml-datasets/inaturalist/v0.1.0/raw/metadata_{year}_{month}.csv
corresponding to when the data was downloaded. In subsequent versions, I
expect we could just have a new metadata file.

This adds over 300,000 audio files compared to the version of
iNaturalist included in esp-data through the AnimalSpeak dataset.