Added hooks to insert datasetinfo into docs (#160)
- Also noticed two warnings when running mkdocs locally so fixed those
too:
  ```
Passing extra options directly under `options` is deprecated. Instead,
pass them under `options.extra`, and update your templates.
  ```

<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":"main"}
```
-->