update wabad dataset (#186)
WABAD was the first dataset I implemented, and as a result it was the
lowest quality implementation. I wanted to updated it since WABAD is
probably the most important dataset we have for detection evals.

Here, I've updated the dataset so that it is more self-contained (i.e.
no weird label mapping logic), and also had the full suite of tests we
want to run (e.g. hashing for the underlying data).