fix for issue 179: a simple "chain" of datasets (#187)
#179 Deals with failure case of dataset concatenation. Basically the
problem is:

1. User creates a concatenated dataset ```ds = ConcatenatedDataset([ds1,
ds2], merge_level="soft")```
2. When iterating, the current version of ConcatenatedDataset assumes
that a key in the source dataset is not transformed into something else
when the source dataset's `__getitem__` is called. The problem occurs
when updating the returned `item` dict from a source dataset with the
same item from the concatenated data backend. If any key is _colliding_
, the item from the concatenated backend overwrites the source.
3. This fails in the case of datasets like `HawaiianBirds` and
`NocturnalBirdMigration` etc, where a field like `selection_table` is a
string in the data backend (i.e. in the csv file), but that string is
converted to a pandas dataframe in the `__getitem__` call. So, in the
ConcatenatedDataset, the backend has concatenated dataframe with the
column `selection_table` being a string, while the returned item from
the source dataset has `type(item["selection_table"]) = pd.DataFrame` ..
the user actually wants the dataframe not the string :(

The reason we were overwriting source with concatenated (point 1) was
because the whole point of dataset concatenation was to allow transforms
like subsampling etc to be applied at a combined dataset level.

The proposed solution (which is not very nice and up for discussion) in
this PR is to add `collision_policy` which allows the user to select the
behavior .

EDIT: `collision_policy` is removed in favor of a second concat type,
called `ChainedDataset` which, as the name suggests, just chains
datasets together while maintaining the Dataset api, which requires
`__getitem__`. This requirement can be dropped and made simpler in
another PR

---------

Co-authored-by: Benjamin Hoffman <72874445+benjaminsshoffman@users.noreply.github.com>
Co-authored-by: Paul Laisné <paul.laisne2004@gmail.com>
Co-authored-by: Milad Alizadeh <686449+mil-ad@users.noreply.github.com>